// Csh.ImageSuite.WebClient/src/app/models/settings.ts
export class TransferJob {
    jobUid: string = "";
    deliveryStatus: string = "";
    patientName: string = "";
    patientId: string = "";
    createdDate: string = "";
    type: string = "";
    imageCount: string = "";
    deliveryName: string = "";
    destination: string = "";
    finishedNumber: string = "";
    successNumber: string = "";
    compressionRatio: string = "";
}

export class TransferJobItem {
    itemNo: string = "";
    type: string = "";
    patientId: string = "";
    name: string = "";
    executeStatus: string = "";
    executeTime: string = "";
}